remove some command stuff and move to Branch
