+++ .idea/misc.xml
@@ -20,6 +20,7 @@
         <option value="$PROJECT_DIR$/core-implementation/pom.xml" />
         <option value="$PROJECT_DIR$/core-omniscient/pom.xml" />
         <option value="$PROJECT_DIR$/generic-impl/pom.xml" />
+        <option value="$PROJECT_DIR$/spigot-examples/pom.xml" />
       </set>
     </option>
   </component>
