+++ core/src/main/java/me/aurium/beetle/api/command/CommandRegisterer.java
@@ -1,9 +1,5 @@
 package me.aurium.beetle.api.command;
 
-import java.util.Collection;
-
 public interface CommandRegisterer<T> {
     void register(CommandRegistry<T> registry);
-
-
 }
