--- core/src/main/java/me/aurium/beetle/api/command/Context.java
@@ -1,12 +1,0 @@
-package me.aurium.beetle.api.command;
-
-public interface Context<T> {
-
-    T getSender();
-    String getAlias();
-    String[] getArgs();
-
-    void messageSender(String string);
-    boolean equals(Object object);
-
-}
