--- core/src/main/java/me/aurium/beetle/api/command/ContextHandler.java
@@ -1,8 +1,0 @@
-package me.aurium.beetle.api.command;
-
-@FunctionalInterface
-public interface ContextHandler<T> {
-
-    boolean handle(Context<T> consumed);
-
-}
