--- core/src/main/java/me/aurium/beetle/api/command/ContextSource.java
@@ -1,7 +1,0 @@
-package me.aurium.beetle.api.command;
-
-public interface ContextSource<T> {
-
-    Context<T> context(T sender, String alias, String[] args);
-
-}
