--- core/src/main/java/me/aurium/beetle/api/command/TabContextHandler.java
@@ -1,10 +1,0 @@
-package me.aurium.beetle.api.command;
-
-import java.util.Collection;
-
-@FunctionalInterface
-public interface TabContextHandler<T> {
-
-    Collection<String> handle(Context<T> context);
-
-}
