+++ pom.xml
@@ -21,7 +21,7 @@
         <module>spigot-impl</module>
         <module>velo-impl</module>
         <module>discord-impl</module>
-        <module>spigot-examples</module>
+        <module>spigot-commons</module>
     </modules>
 
     <build>
