--- spigot-examples/src/main/java/me/aurium/examplebeetle/ExampleConf.java
@@ -1,4 +1,0 @@
-package me.aurium.examplebeetle;
-
-public class ExampleConf {
-}
