+++ spigot-impl/pom.xml
@@ -17,15 +17,8 @@
     <dependencies>
         <dependency>
             <groupId>me.aurium</groupId>
-            <artifactId>beetle-core-default</artifactId>
+            <artifactId>beetle-spigot-commons</artifactId>
             <version>0.4.2-SNAPSHOT</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.spigotmc</groupId>
-            <artifactId>spigot-api</artifactId>
-            <version>1.16.5-R0.1-SNAPSHOT</version>
-            <scope>provided</scope>
         </dependency>
 
         <dependency>
