+++ spigot-impl/src/main/java/me/aurium/beetle/spigot/SpigotStringUtil.java
@@ -5,7 +5,7 @@ import org.bukkit.ChatColor;
 /**
  * Literally shut up i dont care
  */
-public class SpigotString {
+public class SpigotStringUtil {
 
     public static String color(String string) {
         return ChatColor.translateAlternateColorCodes('&', string);
